feat: - combine 3 scanner activity into 1
