--- Staff/src/main/java/com/jaylangkung/brainnet_staff/auth/LoginWebAppActivity.kt
@@ -1,133 +1,0 @@
-package com.jaylangkung.brainnet_staff.auth
-
-import android.content.Intent
-import android.os.Build
-import android.os.Bundle
-import android.os.VibrationEffect
-import android.os.Vibrator
-import android.view.View
-import androidx.activity.OnBackPressedCallback
-import androidx.appcompat.app.AppCompatActivity
-import androidx.core.content.ContextCompat
-import com.budiyev.android.codescanner.*
-import com.jaylangkung.brainnet_staff.MainActivity
-import com.jaylangkung.brainnet_staff.R
-import com.jaylangkung.brainnet_staff.data_class.DefaultResponse
-import com.jaylangkung.brainnet_staff.databinding.ActivityLoginWebAppBinding
-import com.jaylangkung.brainnet_staff.retrofit.DataService
-import com.jaylangkung.brainnet_staff.retrofit.RetrofitClient
-import com.jaylangkung.brainnet_staff.utils.Constants
-import com.jaylangkung.brainnet_staff.utils.ErrorHandler
-import com.jaylangkung.brainnet_staff.utils.MySharedPreferences
-import es.dmoral.toasty.Toasty
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
-
-class LoginWebAppActivity : AppCompatActivity() {
-
-    private lateinit var binding: ActivityLoginWebAppBinding
-    private lateinit var myPreferences: MySharedPreferences
-    private lateinit var codeScanner: CodeScanner
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityLoginWebAppBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-        myPreferences = MySharedPreferences(this@LoginWebAppActivity)
-
-        onBackPressedDispatcher.addCallback(this, object : OnBackPressedCallback(true) {
-            override fun handleOnBackPressed() {
-                startActivity(Intent(this@LoginWebAppActivity, MainActivity::class.java))
-                finish()
-            }
-        })
-
-        val idadmin = myPreferences.getValue(Constants.USER_IDADMIN).toString()
-        val tokenAuth = getString(R.string.token_auth, myPreferences.getValue(Constants.TokenAuth).toString())
-
-        binding.apply {
-            btnBack.setOnClickListener {
-                onBackPressedDispatcher.onBackPressed()
-            }
-
-            codeScanner = CodeScanner(this@LoginWebAppActivity, scannerView).apply {
-                camera = CodeScanner.CAMERA_BACK
-                formats = CodeScanner.ALL_FORMATS
-                autoFocusMode = AutoFocusMode.CONTINUOUS
-                scanMode = ScanMode.SINGLE
-                isAutoFocusEnabled = true
-                isFlashEnabled = false
-                startPreview()
-
-                decodeCallback = DecodeCallback {
-                    runOnUiThread {
-                        binding.loadingAnim.visibility = View.VISIBLE
-                        if (it.text.contains("webapp", ignoreCase = true)) {
-                            insertWebApp(idadmin, it.text, tokenAuth)
-                        } else {
-                            Toasty.warning(this@LoginWebAppActivity, "QR Code tidak valid", Toasty.LENGTH_SHORT, true).show()
-                            vibrate()
-                            onBackPressedDispatcher.onBackPressed()
-                        }
-                    }
-                }
-
-                errorCallback = ErrorCallback {
-                    runOnUiThread {
-                        ErrorHandler().responseHandler(
-                            this@LoginWebAppActivity, "codeScanner | errorCallback", it.message.toString()
-                        )
-                    }
-                }
-            }
-        }
-    }
-
-    override fun onResume() {
-        super.onResume()
-        codeScanner.startPreview()
-    }
-
-    override fun onPause() {
-        codeScanner.releaseResources()
-        super.onPause()
-    }
-
-    private fun vibrate() {
-        val vibrator = ContextCompat.getSystemService(this, Vibrator::class.java) as Vibrator
-        if (Build.VERSION.SDK_INT >= 26) {
-            vibrator.vibrate(VibrationEffect.createOneShot(200, VibrationEffect.DEFAULT_AMPLITUDE))
-        } else {
-            @Suppress("DEPRECATION") vibrator.vibrate(200)
-        }
-    }
-
-    private fun insertWebApp(idadmin: String, device_id: String, tokenAuth: String) {
-        val service = RetrofitClient().apiRequest().create(DataService::class.java)
-        service.insertWebApp(idadmin, device_id, tokenAuth).enqueue(object : Callback<DefaultResponse> {
-            override fun onResponse(call: Call<DefaultResponse>, response: Response<DefaultResponse>) {
-                binding.loadingAnim.visibility = View.GONE
-                if (response.isSuccessful) {
-                    vibrate()
-                    if (response.body()!!.status == "success") {
-                        Toasty.success(this@LoginWebAppActivity, "Login berhasil", Toasty.LENGTH_LONG).show()
-                        onBackPressedDispatcher.onBackPressed()
-                    }
-                } else {
-                    ErrorHandler().responseHandler(
-                        this@LoginWebAppActivity, "insertWebApp | onResponse", response.message()
-                    )
-                }
-            }
-
-            override fun onFailure(call: Call<DefaultResponse>, t: Throwable) {
-                binding.loadingAnim.visibility = View.GONE
-                ErrorHandler().responseHandler(
-                    this@LoginWebAppActivity, "insertWebApp | onResponse", t.message.toString()
-                )
-            }
-        })
-    }
-
-}
